requirements.txt
phidata 
streamlit 
groq
python-dotenv
fastapi